+++ tests/testthat/test-ggsave.R
@@ -37,3 +37,41 @@ test_that("ggsave", {
     fixed = TRUE
   )
 })
+
+test_that("ggsave with all filenames specified (#25)", {
+  d_plot <-
+    data.frame(
+      A = rep(c("foo", "bar"), each = 4),
+      B = 1:8,
+      C = 11:18,
+      Bunit = "mg",
+      Cunit = "km"
+    )
+  all_plots <-
+    ggtibble(
+      d_plot,
+      ggplot2::aes(x = B, y = C),
+      outercols = c("A", "Bunit", "Cunit"),
+      caption = "All the {A}",
+      labs = list(x = "B ({Bunit})", y = "C ({Cunit})")
+    ) +
+    ggplot2::geom_point() +
+    ggplot2::geom_line()
+  expected_files <- file.path(tempdir(), c("foo.png", "bar.png"))
+  expect_equal(
+    ggsave(filename = c("foo.png", "bar.png"), plot = all_plots, path = tempdir()),
+    expected_files
+  )
+  unlink(expected_files)
+
+  expect_error(
+    ggsave(filename = c("foo.png", "bar.png", "baz.png"), plot = all_plots, path = tempdir()),
+    regexp = "Assertion on 'filename' failed: Must have length <= 2, but has length 3.",
+    fixed = TRUE
+  )
+  expect_error(
+    ggsave(filename = character(), plot = all_plots, path = tempdir()),
+    regexp = "Assertion on 'filename' failed: Must have length >= 1, but has length 0.",
+    fixed = TRUE
+  )
+})